requirements.txt
# 数据获取和处理
pandas==1.5.3
numpy==1.24.3
yfinance==0.2.31
akshare==1.9.80
tushare==1.2.89
pytz==2023.3
requests>=2.31.0

# 数据可视化
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.18.0

# Web框架和API
fastapi==0.104.1
uvicorn==0.23.2
gradio==3.50.2
python-multipart==0.0.6

# AI模型集成
openai==1.1.1

# 回测和策略
backtrader==1.9.78.123
pyfolio==0.9.2
alphalens==0.4.0

# 工具库
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.0
schedule==1.2.0

# 兼容性和可选依赖
protobuf<4.0.0
pydantic==1.10.8 